+++ src/k-provider/ott/provider-parser.js
@@ -1,6 +1,8 @@
 //@flow
-import KalturaPlaybackSource from './response-types/kaltura-playback-source'
 import getLogger from '../../util/logger'
+import KalturaPlaybackSource from './response-types/kaltura-playback-source'
+import KalturaPlaybackContext from './response-types/kaltura-playback-context'
+import KalturaAsset from './response-types/kaltura-asset'
 import MediaEntry from '../../entities/media-entry'
 import Drm from '../../entities/drm'
 import MediaSource from '../../entities/media-source'
@@ -9,32 +11,60 @@ import {SupportedStreamFormat} from '../../entities/media-format'
 import KalturaDrmPlaybackPluginData from '../common/response-types/kaltura-drm-playback-plugin-data'
 import BaseProviderParser from '../common/base-provider-parser'
 
+const MediaTypeCombinations: { [mediaType: string]: Object } = {
+  [KalturaAsset.Type.MEDIA]: {
+    [KalturaPlaybackContext.Type.TRAILER]: () => ({type: MediaEntry.Type.VOD}),
+    [KalturaPlaybackContext.Type.PLAYBACK]: (metadata) => {
+      if (metadata.linearAssetId) {
+        return {type: MediaEntry.Type.LIVE, dvrStatus: 0};
+      }
+      return {type: MediaEntry.Type.VOD};
+    }
+  },
+  [KalturaAsset.Type.EPG]: {
+    [KalturaPlaybackContext.Type.CATCHUP]: () => ({type: MediaEntry.Type.VOD}),
+    [KalturaPlaybackContext.Type.START_OVER]: () => ({type: MediaEntry.Type.LIVE, dvrStatus: 1})
+  },
+  [KalturaAsset.Type.RECORDING]: {
+    [KalturaPlaybackContext.Type.PLAYBACK]: () => ({type: MediaEntry.Type.VOD})
+  }
+};
+
 export default class OTTProviderParser extends BaseProviderParser {
   static _logger = getLogger("OTTProviderParser");
 
   /**
-   * Returns parsed media entry by given OTT response objects
+   * Returns parsed media entry by given OTT response objects.
    * @function getMediaEntry
-   * @param {any} assetResponse - The asset response
+   * @param {any} assetResponse - The asset response.
+   * @param {Object} requestData - The request data object.
    * @returns {MediaEntry} - The media entry
    * @static
    * @public
    */
-  static getMediaEntry(assetResponse: any): MediaEntry {
+  static getMediaEntry(assetResponse: any, requestData: Object): MediaEntry {
     const mediaEntry = new MediaEntry();
     const playbackContext = assetResponse.playBackContextResult;
     const mediaData = assetResponse.mediaDataResult;
     const kalturaSources = playbackContext.sources;
     mediaEntry.name = mediaData.name;
     mediaEntry.id = mediaData.id;
+
     const metadata = {};
     metadata.description = mediaData.description;
     metadata.poster = OTTProviderParser._getPoster(mediaData.pictures);
     Object.assign(metadata, mediaData.metas);
     Object.assign(metadata, mediaData.tags);
     mediaEntry.metadata = metadata;
-    mediaEntry.sources = OTTProviderParser._getParsedSources(kalturaSources);
+
+    const filteredKalturaSources = OTTProviderParser._filterSourcesByFormats(kalturaSources, requestData.formats);
+    mediaEntry.sources = OTTProviderParser._getParsedSources(filteredKalturaSources);
+
+    const typeData = OTTProviderParser._getMediaType(mediaData, requestData.mediaType, requestData.contextType);
+    mediaEntry.type = typeData.type;
+    mediaEntry.dvrStatus = typeData.dvrStatus;
     mediaEntry.duration = Math.max.apply(Math, kalturaSources.map(source => source.duration));
+
     return mediaEntry;
   }
 
@@ -48,10 +78,44 @@ export default class OTTProviderParser extends BaseProviderParser {
     if (pictures && pictures.length > 0) {
       const picObj = pictures[0];
       const url = picObj.url;
-      const end = url.indexOf('width');
-      return url.substring(0, end - 1);
+      const regex = /(width|height)\/\d*\/(height|width)\/\d*/g;
+      const end = url.search(regex);
+      if (end > -1) {
+        return url.substring(0, end - 1);
+      }
+      return url;
     }
     return '';
+  }
+
+  /**
+   * Gets the media type (LIVE/VOD)
+   * @param {Object} metadata - The asset metadata.
+   * @param {string} mediaType - The asset media type.
+   * @param {string} contextType - The asset context type.
+   * @returns {Object} - The type data object.
+   * @private
+   */
+  static _getMediaType(metadata: Object, mediaType: string, contextType: string): Object {
+    let typeData = {type: MediaEntry.Type.UNKNOWN};
+    if (typeof MediaTypeCombinations[mediaType][contextType] === 'function') {
+      typeData = MediaTypeCombinations[mediaType][contextType](metadata);
+    }
+    return typeData;
+  }
+
+  /**
+   * Filtered the kalturaSources array by device type.
+   * @param {Array<KalturaPlaybackSource>} kalturaSources - The kaltura sources.
+   * @param {Array<string>} formats - Partner device formats.
+   * @returns {Array<KalturaPlaybackSource>} - Filtered kalturaSources array.
+   * @private
+   */
+  static _filterSourcesByFormats(kalturaSources: Array<KalturaPlaybackSource>, formats: Array<string>): Array<KalturaPlaybackSource> {
+    if (formats.length > 0) {
+      kalturaSources = kalturaSources.filter(source => formats.includes(source.type));
+    }
+    return kalturaSources;
   }
 
   /**